Refactor: Move react component files to new file structure, fix imports
